State `arange` type in `test_buffer_coercion`
Just add typing to `arange` for clarity as Windows and Unix builds of
Python and NumPy often handle integers differently.

+++ numcodecs/tests/test_compat.py
@@ -33,7 +33,7 @@ def test_buffer_coercion():
     bufs = [
         b'adsdasdas',
         bytes(20),
-        np.arange(100),
+        np.arange(100, dtype=np.int64),
         array.array('l', b'qwertyuiqwertyui'),
         mmap.mmap(-1, 10)
     ]
